+++ packages/components/choropleth/src/shared.js
@@ -1,0 +1,12 @@
+/**
+* @module @svizzle/choropleth/src/shared
+*/
+
+export const defaultGeometry = {
+	bottom: 10,
+	left: 10,
+	right: 10,
+	top: 10,
+};
+
+export * as projections from './projections';