+++ packages/tools/geo/index.js
@@ -1,1 +1,2 @@
 export * from './src/modules/geojson';
+export * from './src/modules/topojson';