+++ packages/tools/geo/src/modules/topojson.js
@@ -1,27 +1,12 @@
 /**
-* @module @svizzle/choropleth/utils
+* @module @svizzle/geo/src/modules/topojson
 */
 
 import {feature} from 'topojson-client';
-import {setGeometryPrecision} from '@svizzle/geo';
-
-export const defaultGeometry = {
-	bottom: 10,
-	left: 10,
-	right: 10,
-	top: 10,
-};
-
-export const truncateGeojson = setGeometryPrecision(4);
 
 /**
- * Convert a topojson to a geojson truncating coordinates to precision 4.
- * This is quite specific to `@svizzle/choropleth` to limit the geojson weight down for performance reasons.
- * For a full conversion use:
- * ```
- * import {feature} from 'topojson-client';
- * feature(topojson, topojson.objects[id])
- * ```
+ * Convert a topojson to a geojson.
+ * Ported from /choropleth v0.4.0 to /geo v0.7.0, but without truncating digits.
  *
  * @function
  * @arg {object} topojson - the topojson to be converted to geojson
@@ -61,14 +46,15 @@ export const truncateGeojson = setGeometryPrecision(4);
 			'geometry': {
 				'type': 'Polygon',
 				'coordinates': [
-					[[6.3163,50.4967], ...
+					[[6.3163...,50.4967...],
+					...
 				]
 			}
 		}
 	]
 }
  *
- * @since 0.4.0
+ * @since 0.7.0
  */
 export const topoToGeo = (topojson, id) =>
-	truncateGeojson(feature(topojson, topojson.objects[id]));
+	feature(topojson, topojson.objects[id]);